requirements.txt

## 'requirements.txt'

pandas>=2.2
scikit-learn>=1.4
pyyaml>=6.0
mlflow>=2.14
joblib>=1.3
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
httpx>=0.27
evidently>=0.4.30
dvc[gcs]>=3.50
pytest>=8.2
ruff>=0.5
pre-commit>=3.7
